requirements.txt
pyfaidx
psutil
biopython
pysam
pyarrow
pandas==2.0.3
omegaconf==2.3.0
datasets 
lightning==2.2
wandb
fsspec
scikit-learn
hydra-core==1.3.2
rich==13.7.0
timm==0.9.16
transformers==4.38.1
triton==2.1.0
tqdm==4.66.1
genomic-benchmarks==0.0.9
boto3
s3fs
torch
causal-conv1d>=1.4.0
mamba-ssm
